Usermgmt: Made CreatePermissionLayer public and static
